faster stl export (#4643)
