+++ src/io/export.h
@@ -218,6 +218,7 @@ public:
   bool foreach_vertex(const std::function<bool(const Vertex&)>& callback) const;
   bool foreach_indexed_triangle(const std::function<bool(const std::array<int, 3>&)>& callback) const;
   bool foreach_triangle(const std::function<bool(const std::array<Vertex, 3>&)>& callback) const;
+  void export_to_polyset(PolySet& ps) const;
 
 private:
   std::vector<Vertex> vertices;
