+++ src/io/export_stl.cc
@@ -27,6 +27,7 @@
 #include "export.h"
 #include "PolySet.h"
 #include "PolySetUtils.h"
+#include "double-conversion/double-conversion.h"
 #ifdef ENABLE_MANIFOLD
 #include "ManifoldGeometry.h"
 #endif
@@ -39,121 +40,142 @@
 #endif
 
 namespace {
+/* Define values for double-conversion library. */
+#define DC_BUFFER_SIZE (128)
+#define DC_FLAGS (double_conversion::DoubleToStringConverter::UNIQUE_ZERO)
+#define DC_INF NULL // Only finite values in STL outputs!
+#define DC_NAN NULL
+#define DC_EXP 'e'
+#define DC_DECIMAL_LOW_EXP (-6)
+#define DC_DECIMAL_HIGH_EXP (21)
+#define DC_MAX_LEADING_ZEROES (5)
+#define DC_MAX_TRAILING_ZEROES (0)
 
-std::string toString(const Vector3d& v)
+std::string toString(const Vector3f& v)
 {
-  return STR(v[0], " ", v[1], " ", v[2]);
+  double_conversion::DoubleToStringConverter dc(
+    DC_FLAGS, DC_INF, DC_NAN, DC_EXP,
+    DC_DECIMAL_LOW_EXP, DC_DECIMAL_HIGH_EXP, DC_MAX_LEADING_ZEROES, DC_MAX_TRAILING_ZEROES
+  );
+
+  char buffer[DC_BUFFER_SIZE];
+
+  double_conversion::StringBuilder builder(buffer, DC_BUFFER_SIZE);
+  dc.ToShortestSingle(v[0], &builder);
+  builder.AddCharacter(' ');
+  dc.ToShortestSingle(v[1], &builder);
+  builder.AddCharacter(' ');
+  dc.ToShortestSingle(v[2], &builder);
+  builder.Finalize();
+
+  return buffer;
 }
 
-Vector3d toVector(const std::array<double, 3>& pt) {
-  return {pt[0], pt[1], pt[2]};
+int32_t flipEndianness(int32_t x) {
+  return 
+    ((x << 24) & 0xff000000) | ((x >> 24) & 0xff) |
+    ((x << 8) & 0xff0000) | ((x >> 8) & 0xff00);
 }
 
-Vector3d fromString(const std::string& vertexString)
-{
-  Vector3d v;
-  std::istringstream stream{vertexString};
-  stream >> v[0] >> v[1] >> v[2];
-  return v;
-}
+template <size_t N>
+void write_floats(std::ostream& output, const std::array<float, N>& data) {
+  static uint16_t test = 0x0001;
+  static bool isLittleEndian = *reinterpret_cast<char *>(&test) == 1;
 
-void write_vector(std::ostream& output, const Vector3f& v) {
-  for (int i = 0; i < 3; ++i) {
-    static_assert(sizeof(float) == 4, "Need 32 bit float");
-    float f = v[i];
-    char *fbeg = reinterpret_cast<char *>(&f);
-    char data[4];
+  if (isLittleEndian) {
+    output.write(reinterpret_cast<char *>(const_cast<float *>(&data[0])), N * sizeof(float));
+  } else {
+    std::array<float, N> copy(data);
 
-    uint16_t test = 0x0001;
-    if (*reinterpret_cast<char *>(&test) == 1) {
-      std::copy(fbeg, fbeg + 4, data);
-    } else {
-      std::reverse_copy(fbeg, fbeg + 4, data);
+    auto *ints = reinterpret_cast<int32_t *>(&copy[0]);
+    for (size_t i = 0; i < N; i++) {
+      ints[i] = flipEndianness(ints[i]);
     }
-    output.write(data, 4);
+
+    output.write(reinterpret_cast<char *>(&copy[0]), N * sizeof(float));
   }
 }
 
 
-uint64_t append_stl(const PolySet& ps, std::ostream& output, bool binary)
+uint64_t append_stl(const PolySet& polyset, std::ostream& output, bool binary)
 {
-  uint64_t triangle_count = 0;
-  PolySet triangulated(3);
-  PolySetUtils::tessellate_faces(ps, triangulated);
-
-  auto processTriangle = [&](const std::array<Vector3d, 3>& p) {
-      triangle_count++;
-
-      if (binary) {
-        Vector3f p0 = p[0].cast<float>();
-        Vector3f p1 = p[1].cast<float>();
-        Vector3f p2 = p[2].cast<float>();
-
-        // Ensure 3 distinct vertices.
-        if ((p0 != p1) && (p0 != p2) && (p1 != p2)) {
-          Vector3f normal = (p1 - p0).cross(p2 - p0);
-          normal.normalize();
-          if (!is_finite(normal) || is_nan(normal)) {
-            // Collinear vertices.
-            normal << 0, 0, 0;
-          }
-          write_vector(output, normal);
-        }
-        write_vector(output, p0);
-        write_vector(output, p1);
-        write_vector(output, p2);
-        char attrib[2] = {0, 0};
-        output.write(attrib, 2);
-      } else { // ascii
-        std::array<std::string, 3> vertexStrings;
-        std::transform(p.cbegin(), p.cend(), vertexStrings.begin(),
-                       toString);
-
-        if (vertexStrings[0] != vertexStrings[1] &&
-            vertexStrings[0] != vertexStrings[2] &&
-            vertexStrings[1] != vertexStrings[2]) {
-
-          // The above condition ensures that there are 3 distinct
-          // vertices, but they may be collinear. If they are, the unit
-          // normal is meaningless so the default value of "0 0 0" can
-          // be used. If the vertices are not collinear then the unit
-          // normal must be calculated from the components.
-          output << "  facet normal ";
-
-          Vector3d p0 = fromString(vertexStrings[0]);
-          Vector3d p1 = fromString(vertexStrings[1]);
-          Vector3d p2 = fromString(vertexStrings[2]);
-
-          Vector3d normal = (p1 - p0).cross(p2 - p0);
-          normal.normalize();
-          if (is_finite(normal) && !is_nan(normal)) {
-            output << normal[0] << " " << normal[1] << " " << normal[2]
-                   << "\n";
-          } else {
-            output << "0 0 0\n";
-          }
-          output << "    outer loop\n";
-
-          for (const auto& vertexString : vertexStrings) {
-            output << "      vertex " << vertexString << "\n";
-          }
-          output << "    endloop\n";
-          output << "  endfacet\n";
-        }
-      }
-    };
-
+  static_assert(sizeof(float) == 4, "Need 32 bit float");
+  // check if tessellation is needed
+  PolySet tmp(3);
+  bool needTessellation = !std::all_of(
+      polyset.indices.begin(), polyset.indices.end(),
+      [](auto& face) { return face.size() == 3; }
+  );
+  if (needTessellation)
+    PolySetUtils::tessellate_faces(polyset, tmp);
+  bool useTmp = needTessellation;
   if (Feature::ExperimentalPredictibleOutput.is_enabled()) {
-    Export::ExportMesh exportMesh { triangulated };
-    exportMesh.foreach_triangle([&](const auto& pts) {
-        processTriangle({ toVector(pts[0]), toVector(pts[1]), toVector(pts[2]) });
-        return true;
-      });
-  } else {
-    for (const auto& p : triangulated.indices) {
-      assert(p.size() == 3); // STL only allows triangles
-      processTriangle({ triangulated.vertices[p[0]], triangulated.vertices[p[1]], triangulated.vertices[p[2]] });
+    Export::ExportMesh ex(needTessellation ? tmp : polyset);
+    ex.export_to_polyset(tmp);
+    useTmp = true;
+  }
+  const PolySet& ps = useTmp ? tmp : polyset;
+
+  uint64_t triangle_count = 0;
+
+  // In ASCII mode only, convert each vertex to string.
+  std::vector<std::string> vertexStrings;
+  if (!binary) {
+    vertexStrings.resize(ps.vertices.size());
+    std::transform(ps.vertices.begin(), ps.vertices.end(), vertexStrings.begin(),
+      [](auto &p) { return toString({p.x(), p.y(), p.z()}); });
+  }
+
+  // Used for binary mode only
+  std::array<float, 4lu * 3> coords;
+
+  for (const auto &t : ps.indices) {
+    const auto &p0 = ps.vertices[t[0]];
+    const auto &p1 = ps.vertices[t[1]];
+    const auto &p2 = ps.vertices[t[2]];
+
+    // Tessellation already eliminated these cases.
+    assert(p0 != p1 && p0 != p2 && p1 != p2);
+
+    auto normal = (p1 - p0).cross(p2 - p0);
+    if (!normal.isZero(0)) {
+      normal.normalize();
     }
+
+    if (binary) {
+      auto coords_offset = 0;
+      auto addCoords = [&](auto v) {
+        for (auto i : {0, 1, 2})
+          coords[coords_offset++] = v[i];
+      };
+      addCoords(normal);
+      addCoords(p0);
+      addCoords(p1);
+      addCoords(p2);
+      assert(coords_offset == 4 * 3);
+      write_floats(output, coords);
+      char attrib[2] = {0, 0};
+      output.write(attrib, 2);
+    } else {
+      const auto &s0 = vertexStrings[t[0]];
+      const auto &s1 = vertexStrings[t[1]];
+      const auto &s2 = vertexStrings[t[2]];
+
+      // Since the points are different, the precision we use to 
+      // format them to string should guarantee the strings are 
+      // different too.
+      assert(s0 != s1 && s0 != s2 && s1 != s2);
+      
+      output << "  facet normal ";
+      output << toString({normal.x(), normal.y(), normal.z()}) << "\n";
+      output << "    outer loop\n";
+      output << "      vertex " << s0 << "\n";
+      output << "      vertex " << s1 << "\n";
+      output << "      vertex " << s2 << "\n";
+      output << "    endloop\n";
+      output << "  endfacet\n";
+    }
+    triangle_count++;
   }
 
   return triangle_count;
